Restructured classrooms controller
